Move Various GetVerbEvents to Shared. (#2536)
# Description

This PR moves some GetVerbsEvent subscriptions from Content.Server to
Content.Shared. This means that the verbs in question pop up instantly
on the client, rather than popping in a second after the user has right
clicked on an entity. In other words, the UI should be more consistent.

The _acts_ for these verbs have not been predicted. In the case of
`StackSystem`, this is to avoid predicting interactions with spawned
entities, for `RotateateableSystem`, in order to prevent a visual bug,
and for `SharpSystem` because I just couldn't figure it out.

The three systems mentioned above are the only ones affected by this PR.
I chose RotateableSystem because it was easy, StackSystem because I've
already fixed StackSystem verb prediction on wizden (so I know how to do
it already), and SharpSystem because having the "Butcher" verb pop up
with a delay AND the "Eat" verb pop up with a delay makes the right
click menu almost unusable on clothes as a moth. Unfortunately I tried
my hand at predicting FoodSystem verbs, but I didn't manage to pull it
off.

---

# TODO

- [x] Move some GetVerbEvents to Content.Shared.

---

# Media

See media at
https://github.com/space-wizards/space-station-14/pull/39741 for an
example.

---

# Changelog
