+++ Content.Client/Kitchen/EntitySystems/SharpSystem.cs
@@ -1,0 +1,11 @@
+using Content.Shared.Kitchen.EntitySystems;
+
+namespace Content.Client.Kitchen.EntitySystems;
+
+public sealed class SharpSystem : SharedSharpSystem
+{
+    public override void Initialize()
+    {
+        base.Initialize();
+    }
+}
