+++ Content.Server/Botany/Systems/BotanySystem.Seed.cs
@@ -1,5 +1,5 @@
 using Content.Server.Botany.Components;
-using Content.Server.Kitchen.Components;
+using Content.Shared.Kitchen.Components;
 using Content.Server.Popups;
 using Content.Shared.Chemistry.EntitySystems;
 using Content.Shared.Botany;
