+++ Content.Server/_EE/Supermatter/Systems/SupermatterSystem.cs
@@ -3,7 +3,7 @@ using Content.Server.Atmos.Piping.Components;
 using Content.Server.Chat.Systems;
 using Content.Server.DoAfter;
 using Content.Server.Explosion.EntitySystems;
-using Content.Server.Kitchen.Components;
+using Content.Shared.Kitchen.Components;
 using Content.Server.Lightning;
 using Content.Server.Popups;
 using Content.Server.Radio.EntitySystems;
