+++ Content.Server/_Lavaland/Weapons/WeaponAttachmentSystem.cs
@@ -1,4 +1,4 @@
-using Content.Server.Kitchen.Components;
+using Content.Shared.Kitchen.Components;
 using Content.Shared._Lavaland.Weapons;
 using Content.Shared.Containers.ItemSlots;
 using Content.Shared.Toggleable;
